Replace definition list with unordered list and summaries
