+++ Package.resolved
@@ -6,8 +6,8 @@
         "repositoryURL": "https://github.com/SwiftDocOrg/CommonMark.git",
         "state": {
           "branch": null,
-          "revision": "62176a332884826997928ba578d9d322c16772fb",
-          "version": "0.4.0"
+          "revision": "902cc82abc2e8ad23b73c982eed27c63ae3d9384",
+          "version": null
         }
       },
       {
