+++ Sources/swift-doc/Supporting Types/Pages/HomePage.swift
@@ -63,11 +63,8 @@ struct HomePage: Page {
                 if (!symbols.isEmpty) {
                     Heading { heading }
 
-                    // FIXME: HTMLBlock doesn't render for some reason
-                    Paragraph {
-                        RawHTML {
-                            listHTML(symbols: symbols).description
-                        }
+                    List(of: symbols.sorted()) { symbol in
+                        Abstract(for: symbol).fragment
                     }
                 }
             }
@@ -81,73 +78,21 @@ struct HomePage: Page {
             ("Structures", structures),
             ("Enumerations", enumerations),
             ("Protocols", protocols),
+            ("Typealiases", globalTypealiases),
+            ("Functions", globalFunctions),
+            ("Variables", globalVariables)
         ].compactMap { (heading, symbols) -> HypertextLiteral.HTML? in
             guard !symbols.isEmpty else { return nil }
 
             return #"""
             <section id=\#(heading.lowercased())>
                 <h2>\#(heading)</h2>
-                \#(listHTML(symbols: symbols))
+                <dl>
+                    \#(symbols.sorted().map { Abstract(for: $0).html })
+                </dl>
             </section>
         """#
         })
-        \#(globalsHTML)
-        """#
-    }
-
-    private var globalsHTML: HypertextLiteral.HTML {
-        guard !globalTypealiases.isEmpty ||
-            !globalFunctions.isEmpty ||
-            !globalVariables.isEmpty else {
-                return ""
-        }
-
-        let heading = "Globals"
-        return #"""
-        <section id=\#(heading.lowercased())>
-            <h2>\#(heading)</h2>
-            \#(globalsListHTML)
-        </section>
-        """#
-    }
-
-    private var globalsListHTML: HypertextLiteral.HTML {
-        let globals = [
-            ("Typealiases", globalTypealiases),
-            ("Functions", globalFunctions),
-            ("Variables", globalVariables),
-        ]
-        return #"""
-        \#(globals.compactMap { (heading, symbols) -> HypertextLiteral.HTML? in
-            guard !symbols.isEmpty else { return nil }
-
-            return #"""
-            <section id=\#(heading.lowercased())>
-                <h3>\#(heading)</h3>
-                \#(listHTML(symbols: symbols))
-            </section>
-        """#
-        })
-        """#
-    }
-
-    private func listHTML(symbols: [Symbol]) -> HypertextLiteral.HTML {
-        #"""
-        <dl>
-            \#(symbols.sorted().map { symbol ->  HypertextLiteral.HTML in
-            let descriptor = String(describing: type(of: symbol.api)).lowercased()
-            return #"""
-            <dt class="\#(descriptor)">
-                <a href=\#(path(for: symbol)) title="\#(descriptor) - \#(symbol.id.description)">
-                    \#(softbreak(symbol.id.description))
-                </a>
-            </dt>
-            <dd>
-                \#(commonmark: symbol.documentation?.summary ?? "")
-            </dd>
-            """#
-            })
-        </dl>
         """#
     }
 }
